+++ src/parser/rogue/outlaw/CombatLogParser.js
@@ -14,6 +14,8 @@ import EnergyCapTracker from '../shared/resources/EnergyCapTracker';
 import Energy from './modules/core/Energy';
 import SpellEnergyCost from '../shared/resources/SpellEnergyCost';
 
+import RollTheBonesCastTracker from './modules/core/RollTheBonesCastTracker';
+import RollTheBonesCounter from './modules/core/RollTheBonesCounter';
 import RollTheBonesBuffs from './modules/core/RollTheBonesBuffs';
 import RestlessBlades from './modules/core/RestlessBlades';
 import SliceAndDiceUptime from './modules/talents/SliceAndDiceUptime';
@@ -37,10 +39,13 @@ class CombatLogParser extends CoreCombatLogParser {
 
     //Core
     restlessBlades: RestlessBlades,
-    rolltheBonesBuffs: RollTheBonesBuffs,
+    rollTheBonesCastTracker: RollTheBonesCastTracker,
+
     //Items
 
     //Casts
+    rollTheBonesBuffs: RollTheBonesBuffs,
+    rollTheBonesCounter: rollTheBonesCounter,
 
     //Talents
     sliceAndDiceUptime: SliceAndDiceUptime,
